clippy: remove unnecessary_cast lint allow

+++ crates/tsz-emitter/src/emitter/statements.rs
@@ -379,10 +379,10 @@ impl<'a> Printer<'a> {
                 // won't be emitted again by the end-of-file comment sweep.
                 while self.comment_emit_idx < self.all_comments.len() {
                     let c = &self.all_comments[self.comment_emit_idx];
-                    if c.pos >= comment.pos as u32 && c.end <= comment.end as u32 {
+                    if c.pos >= comment.pos && c.end <= comment.end {
                         self.comment_emit_idx += 1;
                         break;
-                    } else if c.end > comment.end as u32 {
+                    } else if c.end > comment.end {
                         break;
                     }
                     self.comment_emit_idx += 1;
